fix: Move translation out of main so it can be used in testing
